+++ CHANGELOG.md
@@ -1,10 +1,20 @@
-## [0.2.0] - 23-July-2021
+## [0.2.1] - 26-July-2021
 #### Updated by @[Kimman wky](https://github.com/kimmanwky "Kimman wky")
+
+
+* added badge support:
+    * default `MotionBadgeWidget` provided
+    * custom widget allowed
 
 <br/>
 
+
+## [0.2.0] - 23-July-2021
+#### Updated by @[Kimman wky](https://github.com/kimmanwky "Kimman wky")
+
+
 * Update to support null-safety
-* add new parameters:
+* added new parameters:
     * tabBarHeight (double)
     * tabBarColor (Colors)
     * tabSize (double)